+++ documentation/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+February 6, 2019 - bake-toolkit 2.50.0
+    * Added: command line option "--merge-inc" which copies all include files into one temprary folder and invokes the compiler with just one include path.
+    * Added: command line option "--file-cmd" which writes all command line arguments into a file and hands it over to the compiler (with "@" in gcc and dcc).
+    * Added: added Cuda support for GCC toolchain (experimental)
+
 January 23, 2019 - bake-toolkit 2.49.0
     * Added: Support for IAR compiler.